changed price type to int

+++ src/main/java/com/spare4fun/core/entity/Offer.java
@@ -46,7 +46,7 @@ public class Offer implements Serializable {
 
     @Getter
     @Setter
-    private float price;
+    private int price;
 
     @Getter
     @Setter
